Delete talkgroup and system (#69)
Reviewed-on: https://git.dynatron.me/amigan/stillbox/pulls/69
Co-committed-by: Daniel Ponte <amigan@gmail.com>

+++ pkg/talkgroups/xport/export.go
@@ -29,7 +29,7 @@ func (ej *ExportJob) Export(ctx context.Context, w io.Writer) error {
 	var err error
 	tgst := tgstore.FromCtx(ctx)
 	if ej.TalkgroupFilter.IsEmpty() {
-		tgs, err = tgst.SystemTGs(ctx, int32(ej.SystemID))
+		tgs, err = tgst.SystemTGs(ctx, ej.SystemID)
 		if err != nil {
 			return err
 		}
